feat(suites): add AEGP Persistent Data Suite and HQ world transform
Implement the AEGP Persistent Data Suite (v3) as an in-memory preferences
store so aescripts-licensed plugins run their license check to completion
instead of aborting on a missing suite; register it in SUITE_CONTAINER and
the acquire dispatcher. Add the high-quality world-transform convolution
path and refine the handle/world/app/ansi/iterate suites.

+++ aexlo/src/suites/iterate.rs
@@ -167,7 +167,7 @@ pub(super) unsafe extern "C" fn iterate_8_sys(
 				let err = unsafe { func(refcon_ptr, current_x, current_y, src_pixel, dst_pixel) };
 				if err != PF_Err_NONE as PF_Err {
 					// Attempt to store the first error. We don't care if we overwrite another error or lose one race.
-					error_capsule.store(err, Ordering::Relaxed);
+					error_capsule.store(err as u32, Ordering::Relaxed);
 					return; // Stop processing this row
 				}
 			}
